Move code to Catalyst package.
